refactor: rewrite cjsToEsm to it's equivalent in ECMAScript notation
