+++ src/modules/esm.mjs
@@ -1,0 +1,45 @@
+import { dirname, sep } from "path";
+import { release, version } from "os";
+import { createServer } from "http";
+import unknownObjectA from "./files/a.json" assert { type: "json" };
+import unknownObjectB from "./files/b.json" assert { type: "json" };
+import { fileURLToPath } from "url";
+import('./files/c.js')
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = unknownObjectA
+} else {
+    unknownObject = unknownObjectB
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${sep}"`);
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServer((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {
+    unknownObject,
+    myServer,
+};
+
